src/backend/pyproject.toml
[project]
name = "aegis-backend"
version = "0.1.0"
description = "Aegis backend rebuilt as a service-oriented FastAPI application."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "PyJWT[crypto]>=2.12,<3.0"
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.3,<9.0",
  "httpx>=0.27,<1.0",
  "ruff>=0.6,<1.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
